Also appropos from circa Spring 2005 in light of the fantastic new year's festivities here and abroad (so to speak)
N.b., the maintainer also loathsomely looks to reach out to Professor Sam Kamin with thanks for the recent link to the thesis document on which he was the first advisorial coauthor. CS@Illinois Professor Emeritus Pitt on the second line also consented to its meritorious publication -- the fellatio in print and insult to UIUC royalty in the sleighted name of creationism versus annotation it is in its totality. Thank you all for your time and service on the job!

http://slice.cs.uiuc.edu/pubs/peiper_thesis.pdf